2-Multi-Agent-System/requirements.txt
langchain-google-genai
langgraph
python-dotenv
streamlit